requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# News and content extraction
newsapi-python>=0.2.7
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
jieba>=0.42.1

# Machine learning and transformers
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# Visualization and plotting
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Network analysis
networkx>=3.1

# Progress bars and UI
rich>=13.0.0
tqdm>=4.65.0
colorama>=0.4.6

# Image processing
Pillow>=10.0.0

# Utilities
feedparser>=6.0.10
feedfinder2>=0.0.4
langcodes>=3.3.0
